moved common components to shared
