+++ src/components/Layout/Header/Header.tsx
@@ -7,7 +7,7 @@ import { Trans, useTranslation } from "react-i18next";
 import useOnClickOutside from "use-onclickoutside";
 
 import { Connect } from "components/shared/Connect";
-import { TopInfoBar } from "components/TopInfoBar";
+import { TopInfoBar } from "components/shared/TopInfoBar";
 import { useTheme } from "theme";
 
 import { Menu } from "./Menu";
