+++ src/components/Layout/Layout.tsx
@@ -1,7 +1,6 @@
-import { Footer } from "components/Footer";
-import { Header } from "components/Header";
-
 import { Container } from "./Container";
+import { Footer } from "./Footer";
+import { Header } from "./Header";
 import { contentStyles, styles } from "./styles";
 
 type LayoutProps = {
