+++ src/components/Layout/index.ts
@@ -1,1 +1,2 @@
 export { Layout } from "./Layout";
+export { Container } from "./Container";
