+++ src/components/Trade/LiquidityPool/LiquidityPool.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from "react-i18next";
 
+import { Container } from "components/Layout";
 import { DisplayPrice } from "components/shared/DisplayPrice";
-import { Container } from "components/Layout/Container";
 import { useGetLiquidityPoolBalances } from "queries/trade/useGetLiquidityPoolBalances";
 import { useGetLiquidityPools } from "queries/trade/useGetLiquidityPools";
